requirements.txt
urllib3==1.21.1
requests2==2.16.0
pandas==1.3.5
lxml==4.6.5
beautifulsoup4==4.8.2
selenium==3.141.0
